Improved regexp test for pmap
The change in pmap is necessary if it happens that the whole output of
pmap -X or -XX has been done before the second regexp is matched. Since
the matching is greedy, it is matched by the full output, so that
nothing is left for the third regexp and the test fails with
FAIL: extra extended output (footer).

Reference: http://comments.gmane.org/gmane.linux.lfs.devel/13961

Signed-off-by: Craig Small <csmall@enc.com.au>